requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cloudinary-storage==0.3.0
Pillow==9.5.0